♻️ swagger 코드 분리
