+++ app/backend/libs/utils/swagger.ts
@@ -1,0 +1,14 @@
+import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+import { INestApplication } from '@nestjs/common';
+
+export function setupSwagger(app: INestApplication): void {
+  const config = new DocumentBuilder()
+    .setTitle('Morak')
+    .setDescription('Morak API description')
+    .setVersion('1.0')
+    .addTag('moraks')
+    .build();
+
+  const document = SwaggerModule.createDocument(app, config);
+  SwaggerModule.setup('api', app, document);
+}
