+++ app/backend/src/main.ts
@@ -1,23 +1,15 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 import * as cookieParser from 'cookie-parser';
 import { ValidationPipe } from '@nestjs/common';
+import { setupSwagger } from '../libs/utils/swagger';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
   app.use(cookieParser());
-
-  const config = new DocumentBuilder()
-    .setTitle('Morak')
-    .setDescription('Morak API description')
-    .setVersion('1.0')
-    .addTag('moraks')
-    .build();
-  const document = SwaggerModule.createDocument(app, config);
-  SwaggerModule.setup('api', app, document);
+  setupSwagger(app);
 
   app.useGlobalPipes(new ValidationPipe({ whitelist: true, forbidNonWhitelisted: true, transform: true }));
-  await app.listen(8080);
+  await app.listen(process.env.PORT);
 }
 bootstrap();
